+++ Plugins/Picker/Source/NexusPicker/Public/NArcPicker.h
@@ -5,6 +5,7 @@
 
 #include "CoreMinimal.h"
 #include "NArcPickerParams.h"
+#include "Math/NMersenneTwister.h"
 
 /**
  * Provides various functions for generating points as part of an arc using different
@@ -53,6 +54,13 @@ public:
 	 */
 	static void Tracked(TArray<FVector>& OutLocations, int32& Seed, const FNArcPickerParams& Params);
 	
+	/**
+	 * Generate random points as part of an arc using a provided Mersenne Twister.	 
+	 * @param OutLocations An array to store the generated points.
+	 * @param Twister The Mersenne Twister to query for random.
+	 * @param Params The parameters for the point generation.
+	 */
+	static void Twisted(TArray<FVector>& OutLocations, FNMersenneTwister& Twister, const FNArcPickerParams& Params);
 	
 	/**
 	 * Checks if a point is inside or on the surface of the axis-aligned FBox.