+++ Plugins/Picker/Source/NexusPicker/Public/NCirclePicker.h
@@ -5,6 +5,7 @@
 
 #include "CoreMinimal.h"
 #include "NCirclePickerParams.h"
+#include "Math/NMersenneTwister.h"
 
 /**
  * Provides various functions for generating points in the plane of a circle using different
@@ -53,6 +54,14 @@ public:
 	 * @param Params The parameters for the point generation.
 	 */
 	static void Tracked(TArray<FVector>& OutLocations, int32& Seed, const FNCirclePickerParams& Params);
+	
+	/**
+	 * Generate random points inside or on the perimeter of a circle using a provided Mersenne Twister.	 
+	 * @param OutLocations An array to store the generated points.
+	 * @param Twister The Mersenne Twister to query for random.
+	 * @param Params The parameters for the point generation.
+	 */
+	static void Twisted(TArray<FVector>& OutLocations, FNMersenneTwister& Twister, const FNCirclePickerParams& Params);
 	
 	/**
 	 * Checks if a point is inside or on the perimeter of a circle.