+++ Plugins/Picker/Source/NexusPicker/Public/NSpherePicker.h
@@ -5,6 +5,7 @@
 
 #include "CoreMinimal.h"
 #include "NSpherePickerParams.h"
+#include "Math/NMersenneTwister.h"
 
 /** 
  * Provides various functions for generating points inside or on the surface of a sphere using different
@@ -54,6 +55,14 @@ public:
 	 */
 	static void Tracked(TArray<FVector>& OutLocations, int32& Seed, const FNSpherePickerParams& Params);	
 	
+	/**
+	 * Generate random points inside or on the surface of a sphere using a provided Mersenne Twister.	 
+	 * @param OutLocations An array to store the generated points.
+	 * @param Twister The Mersenne Twister to query for random.
+	 * @param Params The parameters for the point generation.
+	 */
+	static void Twisted(TArray<FVector>& OutLocations, FNMersenneTwister& Twister, const FNSpherePickerParams& Params);
+
 	/**
 	 * Checks if a point is inside or on the surface of a sphere.
 	 * @param Origin The center point of the sphere.