+++ Plugins/Picker/Source/NexusPicker/Public/NSplinePicker.h
@@ -55,7 +55,15 @@ public:
 	 * @param Params The parameters for the point generation.
 	 */
 	static void Tracked(TArray<FVector>& OutLocations, int32& Seed, const FNSplinePickerParams& Params);	
-
+	
+	/**
+	 * Generates random points on a spline using a provided Mersenne Twister.	 
+	 * @param OutLocations An array to store the generated points.
+	 * @param Twister The Mersenne Twister to query for random.
+	 * @param Params The parameters for the point generation.
+	 */
+	static void Twisted(TArray<FVector>& OutLocations, FNMersenneTwister& Twister, const FNSplinePickerParams& Params);
+	
 	/**
 	 * Checks if a point is on a spline within a specified tolerance.
 	 * Uses the N_PICKER_TOLERANCE defined in NPickerUtils.h for proximity checking.