+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `FNMultiplayerEditorModule` now has `OnMultiplayerTestStart` and `OnMultiplayerTestEnd` delegates, the first also allows for additional arguments to be passed in to the launch of the client(s)/server.
 - Preference to clean logs folder prior to launching `MultiplayerTest`.
 - `FNDeveloperUtils` and `UNDeveloperLibrary` now have `IsBuildDemo()` methods to support `IS_BUILD_DEMO` definitions from build targets.
+- Pickers all now have a `Twisted` method to support passing a `FNMersenneTwister`.
 
 ### Changed